pyproject.toml
[project]
name = "bitsatards-bot"
version = "2.5.0"
description = "Official project-typesettings for the bitsatards-bot project."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "discord-py>=2.6.4",
    "lxml>=6.0.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.14.14",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.46",
]
